requirements.txt
Flask
Jinja2
numpy
opencv-python
pywebview
sgfmill
